added helper file for new-suggestion tests
