--- cypress/e2e/new-suggestion/review/new-suggestion-review.ts
@@ -1,13 +1,0 @@
-import { Area } from "@/constants/area"
-import { areaInStateShouldBe } from "../new-suggestion"
-
-const instrumentsArea = "area-instruments"
-const areaSongInformation = "area-song-information"
-const toReviewButton = "to-review-button"
-
-export const shouldGoToReviewArea = () => {
-  cy.data(toReviewButton).click()
-  cy.data(instrumentsArea).should("not.exist")
-  cy.data(areaSongInformation).should("not.exist")
-  areaInStateShouldBe(Area.Review)
-}
